frontend/requirements.txt
Flask==3.0.0
Werkzeug>=3.0.3
requests>=2.31.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
# LAB04 dependencies
minio==7.2.0
redis==5.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Pillow>=10.3.0
# Testing
pytest>=7.0.0
